backend/requirements.txt
psycopg2-binary==2.9.9
SQLAlchemy==2.0.27
alembic==1.13.1
python-dotenv==1.0.0
flask==3.0.2
flask-cors==4.0.0

